Docs: Add copy animation and show snackbar #871 show snackbar finished
