+++ src/routes/docs/+layout.svelte
@@ -1,12 +1,15 @@
 <script lang="ts">
+	import { afterNavigate } from '$app/navigation'
 	import Navbar from '$lib/components/navbar.svelte'
-	import SideBar from './side-bar.svelte'
-	import SearchModale from './search-modale.svelte'
+	import Snackbar from '$lib/components/snackbar.svelte'
+	import { KeyboardShortcutHandler } from '$lib/view/keyboard_shortcut_handler'
+	import { WebLogger } from '$lib/view/log/web_logger'
+	import { onMount } from 'svelte'
+	import { _ } from 'svelte-i18n'
 	import MobileSideBar from './mobile-side-bar.svelte'
 	import NavbarSecondRow from './navbar-second-row.svelte'
-	import { afterNavigate } from '$app/navigation'
-	import { onMount } from 'svelte'
-	import { KeyboardShortcutHandler } from '$lib/view/keyboard_shortcut_handler'
+	import SearchModale from './search-modale.svelte'
+	import SideBar from './side-bar.svelte'
 
 	let search_modale_open = false
 	let mobile_side_bar_open = false
@@ -16,6 +19,10 @@
 	$: sections = data?.sections ?? []
 
 	let search_query = ''
+	let copied_snackbar_visible = false
+	let copied_snackbar_timeout: number | undefined
+
+	const web_logger = new WebLogger('docs')
 
 	function open_search_modale(): void {
 		const selection = window.getSelection()
@@ -58,8 +65,42 @@
 		open_search_modale()
 	}
 
+	function add_copy_code_event(): void {
+		const copy_code_elements = document.querySelectorAll('.copy-code')
+
+		copy_code_elements.forEach((element) => {
+			element.addEventListener('click', (event) => {
+				if (!event.target) return
+
+				const target_element = event.target as HTMLElement
+				let current_element = target_element.parentElement
+
+				while (current_element) {
+					if (current_element.classList.contains('code-container')) break
+
+					current_element = current_element.parentElement
+				}
+
+				if (!current_element) return
+
+				const code = current_element.querySelector('code')?.textContent ?? ''
+
+				navigator.clipboard.writeText(code)
+				web_logger.info('on_copy: ' + code)
+
+				if (copied_snackbar_timeout) clearTimeout(copied_snackbar_timeout)
+
+				copied_snackbar_visible = true
+				copied_snackbar_timeout = window.setTimeout(() => {
+					copied_snackbar_visible = false
+				}, 2000)
+			})
+		})
+	}
+
 	afterNavigate(() => {
 		close_mobile_side_bar()
+		add_copy_code_event()
 	})
 
 	onMount(() => {
@@ -89,6 +130,8 @@
 			<slot />
 		</div>
 	</div>
+
+	<Snackbar text={$_('copied')} visible={copied_snackbar_visible} />
 </div>
 
 <style lang="postcss">
